+++ src/aks-preview/HISTORY.rst
@@ -11,6 +11,11 @@ To release a new version, please select a new version number (usually plus 1 to 
 
 Pending
 +++++++
+* Raise InvalidArgumentValueError for azure cni + pod_cidr without overlay.
+
+0.5.126
++++++++
+* Add `--nrg-lockdown-restriction-level <restriction level>` option for chosing the node resource group restriction level in `aks create` and `aks update`
 
 0.5.125
 +++++++